+++ CHANGELOG.md
@@ -1,4 +1,21 @@
-0.87.0 Release notes (YYYY-MM-DD)
+x.x.x Release notes (yyyy-MM-dd)
+=============================================================
+
+### API breaking changes
+
+* None.
+
+### Enhancements
+
+* None.
+
+### Bugfixes
+
+* Throw an exception when passing an array containing a
+  non-RLMObject to -[RLMRealm addObjects:].
+
+
+0.87.0 Release notes (2014-10-21)
 =============================================================
 
 ### API breaking changes
@@ -15,6 +32,7 @@
 * `arraySortedByProperty:ascending:` was renamed to `sortedResultsUsingProperty:ascending`
 * `addObjectsFromArray:` on both `RLMRealm` and `RLMArray` has been renamed to `addObjects:`
   and now accepts any container class which implements `NSFastEnumeration`
+* Building with Swift support now requires Xcode 6.1
 
 ### Enhancements
 
@@ -28,6 +46,7 @@
 * Make creating `RLMRealm` instances on background threads when an instance
   exists on another thread take a fifth of the time.
 * Support for partial updates when calling `createOrUpdateWithObject:` and `addOrUpdateObject:`
+* Re-enable Swift support on OS X
 
 ### Bugfixes